+++ src/Services/CRM/CRMServiceBuilder.php
@@ -72,6 +72,18 @@ class CRMServiceBuilder extends AbstractServiceBuilder
         return $this->serviceCache[__METHOD__];
     }
     
+    public function leadDetailsConfiguration(): Lead\Service\LeadDetailsConfiguration
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new Lead\Service\LeadDetailsConfiguration(
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
     public function contactDetailsConfiguration(): Contact\Service\ContactDetailsConfiguration
     {
         if (!isset($this->serviceCache[__METHOD__])) {